backend/requirements.txt
pandas
numpy
scikit-learn
joblib
python-dotenv
cerebras-cloud-sdk